Refactored foo.a headers for clarity.
