+++ src/foo/bar.cpp
@@ -1,4 +1,4 @@
-#include "foo.h"
+#include "bar.h"
 
 // using namespace foo;
