+++ src/foo/bar.h
