+++ test/unit/foo/bar.cpp
@@ -1,5 +1,5 @@
 #include "catch2/catch.hpp"
-#include "foo.h"
+#include "bar.h"
 
 TEST_CASE("The answer to life, the universe and everything", "[towels]") {
     REQUIRE(foo::bar() == 42);
